Rollback to previous project structure. Refactor Home Page.
